[SCB-2801]remove unused code for new filter
