--- common/common-rest/src/test/java/org/apache/servicecomb/common/rest/filter/TestHttpClientFilter.java
@@ -1,33 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.servicecomb.common.rest.filter;
-
-import java.util.Arrays;
-import java.util.concurrent.ExecutionException;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class TestHttpClientFilter {
-  @Test
-  public void asyncFailed() {
-    HttpClientFilter filter = new HttpClientFilterBaseForTest();
-    HttpClientFilterBeforeSendRequestExecutor executor =
-        new HttpClientFilterBeforeSendRequestExecutor(Arrays.asList(filter), null, null);
-    Assertions.assertThrows(ExecutionException.class, () -> executor.run().get());
-  }
-}
