+++ edge/edge-core/src/main/java/org/apache/servicecomb/edge/core/EdgeAddHeaderFilter.java
@@ -16,22 +16,63 @@
  */
 package org.apache.servicecomb.edge.core;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
+import java.util.function.BiConsumer;
 
 import javax.annotation.Nonnull;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.servicecomb.core.Invocation;
 import org.apache.servicecomb.core.filter.ConsumerFilter;
 import org.apache.servicecomb.core.filter.Filter;
 import org.apache.servicecomb.core.filter.FilterNode;
+import org.apache.servicecomb.foundation.vertx.http.HttpServletRequestEx;
 import org.apache.servicecomb.swagger.invocation.InvocationType;
 import org.apache.servicecomb.swagger.invocation.Response;
 import org.apache.servicecomb.transport.rest.client.RestClientTransportContext;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.netflix.config.ConfigurationManager;
+import com.netflix.config.DynamicPropertyFactory;
 
 public class EdgeAddHeaderFilter implements ConsumerFilter {
+  private static final Logger LOGGER = LoggerFactory.getLogger(EdgeAddHeaderFilter.class);
+
   public static final String NAME = "edge-add-headers";
 
-  private final EdgeAddHeaderClientFilter filter = new EdgeAddHeaderClientFilter();
+  private static final String PREFIX = "servicecomb.edge.filter.addHeader";
+
+  private static final String KEY_ENABLED = PREFIX + ".enabled";
+
+  private static final String KEY_HEADERS = PREFIX + ".allowedHeaders";
+
+  private List<String> publicHeaders = new ArrayList<>();
+
+  private boolean enabled = false;
+
+  public EdgeAddHeaderFilter() {
+    init();
+    ConfigurationManager.getConfigInstance()
+        .addConfigurationListener(event -> {
+          if (StringUtils.startsWith(event.getPropertyName(), PREFIX)) {
+            LOGGER.info("Public headers config have been changed. Event=" + event.getType());
+            init();
+          }
+        });
+  }
+
+  private void init() {
+    enabled = DynamicPropertyFactory.getInstance().getBooleanProperty(KEY_ENABLED, false).get();
+    String publicHeaderStr = DynamicPropertyFactory.getInstance().getStringProperty(KEY_HEADERS, "").get();
+    String[] split = publicHeaderStr.split(",");
+    if (split.length > 0) {
+      publicHeaders = Arrays.asList(split);
+    }
+  }
 
   @Nonnull
   @Override
@@ -41,7 +82,7 @@ public class EdgeAddHeaderFilter implements ConsumerFilter {
 
   @Override
   public boolean isEnabledForTransport(String transport) {
-    return filter.enabled();
+    return enabled;
   }
 
   @Override
@@ -54,10 +95,24 @@ public class EdgeAddHeaderFilter implements ConsumerFilter {
     RestClientTransportContext transportContext = invocation.getTransportContext();
     if (transportContext != null) {
       return CompletableFuture.completedFuture(null)
-          .thenAccept(v -> filter.addHeaders(invocation, transportContext.getHttpClientRequest()::putHeader))
+          .thenAccept(v -> addHeaders(invocation, transportContext.getHttpClientRequest()::putHeader))
           .thenCompose(v -> nextNode.onFilter(invocation));
     }
     // normal consumer in edge process
     return nextNode.onFilter(invocation);
+  }
+
+  public void addHeaders(Invocation invocation, BiConsumer<String, String> headerAdder) {
+    if (!invocation.isEdge()) {
+      return;
+    }
+
+    HttpServletRequestEx oldRequest = invocation.getRequestEx();
+    publicHeaders.forEach(key -> {
+      String value = oldRequest.getHeader(key);
+      if (StringUtils.isNotEmpty(value)) {
+        headerAdder.accept(key, value);
+      }
+    });
   }
 }
